:bug: Correction du problème d'export de la liste des projets filtrées
