+++ src/views/components/buttons/LinkButton.tsx
@@ -1,0 +1,24 @@
+import React from 'react'
+
+type LinkButtonProps = {
+  href: string
+  className?: string
+  children?: React.ReactNode
+  primary?: true
+  download?: true
+}
+
+export const LinkButton = ({ href, className, children, primary, download }: LinkButtonProps) => (
+  <a
+    className={`no-underline inline-flex items-center px-6 py-3 border border-solid text-base font-medium shadow-sm focus:outline-none focus:ring-2 focus:ring-offset-2 ${
+      primary
+        ? 'border-transparent text-white bg-blue-france-sun-base hover:bg-blue-france-sun-hover focus:bg-blue-france-sun-active'
+        : 'border-blue-france-sun-base text-blue-france-sun-base bg-white hover:bg-blue-france-975-base focus:bg-blue-france-975-base'
+    } ${className}`}
+    style={{ color: primary ? 'white' : '#000091', textDecoration: 'none' }}
+    href={href}
+    {...(download && { download })}
+  >
+    {children}
+  </a>
+)
